+++ SistemaCOD/back/src/main/java/com/example/SistemaCOD/repository/DespesaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DespesaRepository extends JpaRepository<Despesa, Integer> {
+public interface DespesaRepository extends JpaRepository<Despesa, Long> {
 
 }
